refactor: move image extraction in in cel/compiler (#13006)
