+++ pkg/imageverification/evaluator/policy.go
@@ -46,7 +46,7 @@ type compiledPolicy struct {
 	matchConditions      []cel.Program
 	matchImageReferences []engine.MatchImageReference
 	validations          []engine.Validation
-	imageExtractors      map[string]variables.CompiledImageExtractor
+	imageExtractors      map[string]engine.ImageExtractor
 	attestors            []*variables.CompiledAttestor
 	attestationList      map[string]string
 	auditAnnotations     map[string]cel.Program
@@ -121,7 +121,7 @@ func (c *compiledPolicy) Evaluate(ctx context.Context, ictx imagedataloader.Imag
 	} else {
 		data[engine.ObjectKey] = request
 	}
-	images, err := variables.ExtractImages(data, c.imageExtractors)
+	images, err := engine.ExtractImages(data, c.imageExtractors)
 	if err != nil {
 		return nil, err
 	}
